backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.4.0
django-filter==23.4
dj-rest-auth
drf-spectacular==0.26.5
psycopg2-binary==2.9.9
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.13.0
azure-storage-blob==12.22.0
Pillow==10.4.0
python-dotenv==1.0.1
requests==2.31.0